Import directly from config package in Vite plugin
